requirements.txt
ultralytics>=8.0
supervision>=0.18
opencv-python>=4.8
matplotlib>=3.7
numpy>=1.24
requests>=2.28
torch>=2.0
torchvision>=0.15
transformers>=4.30
Pillow>=9.0
onnx>=1.14
